moved poll command
